// src/app/modules/supplier/supplier.type.ts
export type ISupplierFilterRequest = {
  searchTerm?: string | undefined
  firstName?: string | undefined
  middleName?: string | undefined
  lastName?: string | undefined
  uniqueId?: string | undefined
  role?: string | undefined
  nid?: string | undefined
  email?: string | undefined
  phoneNo?: string | undefined
  gender?: string | undefined
  presentAddress?: string | undefined
  permanentAddress?: string | undefined
}